+++ gdrive_sync/api.py
@@ -4,9 +4,8 @@ import json
 import logging
 import os
 from datetime import datetime
-from functools import reduce
 from pathlib import Path
-from typing import Dict, Iterable, Iterator, Optional, Tuple
+from typing import Dict, Iterable, Optional
 
 import PyPDF2
 from botocore.exceptions import ClientError
@@ -44,10 +43,9 @@ from websites.constants import (
     RESOURCE_TYPE_IMAGE,
     RESOURCE_TYPE_OTHER,
     RESOURCE_TYPE_VIDEO,
-    WebsiteStarterStatus,
 )
-from websites.models import Website, WebsiteContent, WebsiteStarter
-from websites.site_config_api import ConfigField, ConfigItem, SiteConfig
+from websites.models import Website, WebsiteContent
+from websites.site_config_api import SiteConfig
 from websites.utils import get_valid_base_filename
 
 
@@ -558,130 +556,6 @@ def update_sync_status(website: Website, sync_datetime: datetime):
     website.save()
 
 
-def __iter_all_config_items(website: Website) -> Iterator[Tuple[bool, ConfigItem]]:
-    """
-    Yields ConfigItem for all active starters.
-
-    Args:
-        website (Website): The website being scanned.
-
-    Yields:
-        Iterator[Tuple[bool, ConfigItem]]: A generator where yield[0] indicates whether or not the ConfigItem
-            yield[1] belongs to the starter of `website`.
-    """
-    # Assume the starter of websites without a starter to be the course starter.
-    website_starter_slug = (
-        website.starter.slug if website.starter else settings.OCW_COURSE_STARTER_SLUG
-    )
-    all_starters = WebsiteStarter.objects.filter(
-        status__in=WebsiteStarterStatus.ALLOWED_STATUSES
-    )
-
-    for starter in all_starters:
-        for item in SiteConfig(starter.config).iter_items():
-            yield website_starter_slug == starter.slug, item
-
-
-def __iter_related_fields(
-    item: ConfigItem, only_cross_site: bool
-) -> Iterator[ConfigField]:
-    """
-    Yields ConfigField for each field in `item`.
-
-    Args:
-        item (ConfigItem): The ConfigItem being scanned.
-        only_cross_site (bool): Whether or not to yield only cross site fields.
-
-    Yields:
-        Iterator[ConfigField]: A generator that yields ConfigField.
-    """
-    for field in item.fields:
-        if not only_cross_site or field.get("cross_site", False):
-            yield ConfigField(field)
-
-
-def __get_field_filter(field: dict, resource_id: str, website: Website) -> Optional[Q]:
-    """
-    Generates an appropriate Q expression to filter a field for a resource usage.
-    """
-    q = None
-
-    if field.get("widget") == "markdown" and (
-        CONTENT_TYPE_RESOURCE in field.get("link", [])
-        or CONTENT_TYPE_RESOURCE in field.get("embed", [])
-    ):
-        q = Q(markdown__icontains=resource_id)
-    elif (
-        field.get("widget") == "relation"
-        and field.get("collection") == CONTENT_TYPE_RESOURCE
-    ):
-        lookup_args = ["metadata", field["name"], "content"]
-
-        if field.get("multiple", False):
-            lookup_args.append("contains")
-
-        lookup = "__".join(lookup_args)
-
-        value = (
-            resource_id
-            if not field.get("cross_site", False)
-            else [[resource_id, website.url_path]]
-        )
-
-        q = Q(**{lookup: value})
-    elif field.get("widget") == "menu":
-        lookup = "__".join(("metadata", field.get("name"), "contains"))
-        q = Q(**{lookup: [{"identifier": resource_id}]})
-
-    return q
-
-
-def _get_content_dependencies(drive_file: DriveFile) -> Iterable[WebsiteContent]:
-    """
-    Find and return WebsiteContent that make use of `drive_file` through
-    `drive_file.resource`.
-
-    Args:
-        drive_file (DriveFile): A DriveFile whose dependencies are to be found.
-
-    Returns:
-        Iterable[WebsiteContent]: A list of WebsiteContent that makes use of `drive_file`
-            directly/indirectly.
-    """
-    if drive_file.resource is None:
-        return []
-
-    website = drive_file.website
-    resource_id = drive_file.resource.text_id
-
-    filters = []
-
-    for is_website_config, config_item in __iter_all_config_items(website):
-        for config_field in __iter_related_fields(
-            config_item, only_cross_site=not is_website_config
-        ):
-            field = config_field.field
-            field_q = __get_field_filter(field, resource_id, website)
-
-            if field_q is None:
-                continue
-
-            q = Q(type=config_item.name) & field_q
-
-            if not field.get("cross_site", False):
-                q = Q(website=website) & q
-
-            filters.append(q)
-
-    if not filters:
-        return []
-
-    query = reduce(lambda x, y: x | y, filters)
-    dependencies = WebsiteContent.objects.filter(query)
-
-    return list(dependencies)
-
-
 @transaction.atomic
 def rename_file(obj_text_id, obj_new_filename):
     """Rename the file on S3 associated with the WebsiteContent object to a new filename."""
@@ -703,7 +577,7 @@ def rename_file(obj_text_id, obj_new_filename):
         old_obj = existing_obj.first()
         if old_obj == obj:
             raise Exception("New filename is the same as the existing filename.")
-        dependencies = _get_content_dependencies(old_obj)
+        dependencies = old_obj.get_content_dependencies()
         if dependencies:
             raise Exception(
                 "Not renaming file due to dependencies in existing content: "
@@ -753,7 +627,7 @@ def delete_drive_file(drive_file: DriveFile, sync_datetime: datetime):
     Args:
         drive_file (DriveFile): A drive file.
     """
-    dependencies = _get_content_dependencies(drive_file)
+    dependencies = drive_file.get_content_dependencies()
 
     if dependencies:
         error_message = f"Cannot delete file {drive_file} because it is being used by {dependencies}."
